feat: GetAllList

+++ models/response_model.go
@@ -4,10 +4,10 @@ import "time"
 
 // Pagination 分页结构
 type Pagination struct {
-	CurrentPage int64 `json:"currentPage"`
-	PageSize    int64 `json:"pageSize"`
-	Total       int64 `json:"total"`
-	TotalPage   int64 `json:"totalPage"`
+	CurrentPage int `json:"currentPage"`
+	PageSize    int `json:"pageSize"`
+	Total       int `json:"total"`
+	TotalPage   int `json:"totalPage"`
 }
 
 // ArticlePageResponse 文章分页响应结构
